% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_codes.R
\name{nomis_codes}
\alias{nomis_codes}
\title{Variable codes}
\usage{
nomis_codes(id, code)
}
\arguments{
\item{id}{The ID of the particular dataset. If both \code{id} and 
\code{code} are left empty, returns all available codes for all datasets.}

\item{code}{The variable name to return options for. If left empty, 
returns all options for the dataset specified by \code{id}.}
}
\value{
A list of options.
}
\description{
Retrieve all code options of all datasets, code options for a given 
dataset, or the all the options for a given code variable from a 
particular dataset.
}
\examples{
\dontrun{

z <- nomis_codes("NM_7_1", "geography")

}

}
