% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{nomis_data_access}
\alias{nomis_data_access}
\title{Retrieve nomis datasets}
\usage{
nomis_data_access(id)
}
\arguments{
\item{id}{The ID of the dataset to retrieve.}
}
\value{
A data frame
}
\description{
Retrieve nomis datasets
}
\examples{
\\donotrun{


}
}
