% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\name{nomis_data_info}
\alias{nomis_data_info}
\title{Discovering data availability}
\usage{
nomis_data_info(id)
}
\arguments{
\item{id}{Dataset ID. If empty, returns data on all available datasets.}
}
\value{
A tibble
}
\description{
Retrieve all available data sets or the information available in a specific 
dataset based on its ID.
}
\examples{
\dontrun{
x <- nomis_data_info()
}
}
