% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_metadata.R
\name{nomis_get_metadata}
\alias{nomis_get_metadata}
\title{Nomis metadata concepts and types}
\usage{
nomis_get_metadata(id, concept = NULL, type = NULL,
  additional_queries = NULL)
}
\arguments{
\item{id}{The ID of the particular dataset. Returns no data if not specified.}

\item{concept}{A string with the variable concept to return options for. If 
left empty, returns all the variables for the dataset specified by \code{id}.
Codes are not case sensitive. Defaults to \code{NULL}.}

\item{type}{A string with options for a particular code value, to return
types of variables available for a given code. Defaults to \code{NULL}. If 
\code{concept == NULL}, \code{type} will be ignored.}

\item{additional_queries}{Any other additional queries to pass to the API.
See \url{https://www.nomisweb.co.uk/api/v01/help} for instructions on
query structure. Defaults to \code{NULL}.}
}
\value{
A tibble with options.
}
\description{
Retrieve all code options of all Nomis datasets, code options for a given
dataset, or the all the options for a given code variable from a
particular dataset. Specifying \code{code} will return all the options for a 
given variable in a particular dataset.
}
\examples{
\donttest{

a <- nomis_get_metadata('NM_1_1')

b <- nomis_get_metadata('NM_1_1', 'geography')

c <- nomis_get_metadata('NM_1_1', 'geography', 'TYPE')
#returns all types of geography

d <- nomis_get_metadata('NM_1_1', 'geography', '1879048226')
# returns geography types available within Wigan

}
}
