% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomisr-package.R
\docType{package}
\name{nmisr}
\alias{nmisr}
\alias{nmisr-package}
\title{nomisr: Access nomis data on UK labour market statistics with R
Documentation available at \url{https://www.nomisweb.co.uk/api/v01/help}}
\description{
nomisr: Access nomis data on UK labour market statistics with R
Documentation available at \url{https://www.nomisweb.co.uk/api/v01/help}
}
