% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomisr-package.R
\docType{package}
\name{nmisr}
\alias{nmisr}
\alias{nmisr-package}
\title{nomisr: Access Nomis UK Labour Market Data with R}
\description{
Access UK official statistics from the Nomis database through R. Nomis 
includes data from the Census, the Labour Force Survey, DWP benefit
statistics and other economic and demographic data from the Office for 
National Statistics.
}
\details{
The package provides functions to find what data is available, the 
variables and query options for different datasets and a function for 
downloading data.

Full API documentation available at 
\url{https://www.nomisweb.co.uk/api/v01/help}
}
