% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_codes.R
\name{nomis_codes}
\alias{nomis_codes}
\title{Variable codes}
\usage{
nomis_codes(id, code, type)
}
\arguments{
\item{id}{The ID of the particular dataset. If both \code{id} and 
\code{code} are left empty, returns all available codes for all datasets.}

\item{code}{A string with the variable code to return options for. If left 
empty, returns all the variables for the dataset specified by \code{id}. 
Codes are not case sensitive.}

\item{type}{A string with options for a particular code value, to return 
types of variables available for a given code.}
}
\value{
A list of options.
}
\description{
Retrieve all code options of all datasets, code options for a given 
dataset, or the all the options for a given code variable from a 
particular dataset.
}
\details{
Specifying `code` will return all the options for a given variable.
}
\examples{
\dontrun{

a <- nomis_codes("NM_1_1")

b <- nomis_codes("NM_1_1", "geography")

c <- nomis_codes("NM_1_1", "geography", "TYPE")
#returns all types of geography

d <- nomis_codes("NM_1_1", "geography", "1879048226")
# returns geography types available in Wigan

}

}
