% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_codes.R
\name{nomis_codes}
\alias{nomis_codes}
\title{Variable codes}
\usage{
nomis_codes(id, code = NULL, type = NULL)
}
\arguments{
\item{id}{The ID of the particular dataset. Returns no data if not specified.}

\item{code}{A string with the variable code to return options for. If left 
empty, returns all the variables for the dataset specified by \code{id}. 
Codes are not case sensitive. Defaults to \code{NULL}.}

\item{type}{A string with options for a particular code value, to return 
types of variables available for a given code. Defaults to \code{NULL}.}
}
\value{
A tibble with options.
}
\description{
Retrieve all code options of all Nomis datasets, code options for a given 
dataset, or the all the options for a given code variable from a 
particular dataset.
}
\details{
Specifying \code{code} will return all the options for a given variable.
}
\examples{
\donttest{

a <- nomis_codes('NM_1_1')

b <- nomis_codes('NM_1_1', 'geography')

c <- nomis_codes('NM_1_1', 'geography', 'TYPE')
#returns all types of geography

d <- nomis_codes('NM_1_1', 'geography', '1879048226')
# returns geography types available within Wigan

}

}
