% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{nomis_get_data}
\alias{nomis_get_data}
\title{Retrieve nomis datasets}
\usage{
nomis_get_data(id)
}
\arguments{
\item{id}{The ID of the dataset to retrieve.}
}
\value{
A data frame
}
\description{
Retrieve nomis datasets
}
\examples{
\\donotrun{

x <- nomis_get_data()

}
}
