% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{nomis_search}
\alias{nomis_search}
\title{Search nomis}
\usage{
nomis_search(search, keywords = FALSE)
}
\arguments{
\item{search}{The term to search for.}

\item{keywords}{If \code{TRUE}, searches only in dataset keywords. Defaults 
to \code{FALSE}.}
}
\value{
A tibble with the search results.
}
\description{
Function for searching for datasets on given topics
}
\examples{
\\donotrun{

}
}
