% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{nomis_search}
\alias{nomis_search}
\title{Search nomis datasets}
\usage{
nomis_search(search, keywords = FALSE)
}
\arguments{
\item{search}{A string to search for in the name and description of 
datasets.}

\item{keywords}{If \code{TRUE}, searches in dataset keywords instead of 
names and descriptions. Defaults to \code{FALSE}.}
}
\value{
A tibble with details on all datasets matching the search query.
}
\description{
A function to search for datasets on given topics. The search string by 
default looks in the names and descriptions of datasets.
}
\examples{
\donttest{

x <- nomis_search('seekers')

y <- nomis_search('Claimants', keywords=TRUE)

}
}
